+++ backend/src/modules/pet-types/pet-types.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { PetTypesController } from './pet-types.controller';
+import { PetTypesService } from './pet-types.service';
+
+@Module({
+  controllers: [PetTypesController],
+  providers: [PetTypesService],
+  exports: [PetTypesService],
+})
+export class PetTypesModule {}